add first stab at proper linux-md (software raid) handling
We introduce a new class called GduActivatableDrive (which is even
harder to spell/type than GduPresentable. Sigh.) which is a subclass
of GduDrive. When we see components (e.g. /dev/sda1) we instantiate an
object of this class and when the array appears (e.g. /dev/md0) we set
the device pointer.

This way we'll get a presentable for the raid drive exactly when one
or more components are shown. Haven't hooked the 'start' and 'stop'
actions up yet (requires methods on the daemon) but that should be
rather trivial.

Activatable drives have their own page in GduPageDrive; right now it's
rather empty but the idea is to list the components and, for creation
(plus spare management), allow users to drag components from the tree
into the list.

So lots of work left but this is a good start.