rework device detection to be more robust and rework how Linux MD is handled
Rather than relying on add, remove signals, recompute the full set of
presentables we want. Then diff it against the presentables we already
have and emit the events in the right order. This is significantly
more robust compared to all the special casing we had earlier.

Also fold main functionality of GduActivatableDrive into GduDrive (a
GduDrive now has the notion of whether it is running, ops for
starting/stopping and so forth) and rename the GduActivatableDrive to
GduLinuxMdDrive.