handle inactive and partitioned md devices
See

 http://cgit.freedesktop.org/DeviceKit/DeviceKit-disks/commit/?id=1c6e690dcdc5f31d69d39fc89a9db035802febcc

for the 'inactive' bits. Basically we need this to cope with the current way
autoassembly is handled by Fedora. E.g. if you put in half a mirror and then
remove it again, there's an inactive array with a dangling symlink in sysfs.

For partitioned md devices, this was introduced in the kernel with
this commit

http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=92850bbd71228730c80efd491e7427650188d359

Note that there's a bug in current udev (version 140ish) that prevents
probing the partions of a md device. This will get fixed soon.

To make it possible to partition an md device, we have moved the
"Create partition table" section so it's only visible for a completely
unrecognized device (e.g. when we only see a top-level block device
with no partitions or no recognizable signature).

The move of "Create partition table" makes the UI look a bit empty for
disks that don't report SMART status; there's no sections at all. It
also makes it a bit harder to completely format a disk. The former
should be fixed by having some graphical representation of the
partitions/content, the latter by having a File->New menu that
launches a wizard to prepare formatting the disk.