bug 576431 – better warning
I was trying to delete a partition that was inside an extended partition. The
partitions I have are

part1
part2
part3
[extended]
  part4
  part5
  part6

I was going to delete part6. I got a warning dialog that was talking about
deleting "partition 7 on the media". This had me slightly confused, because my
naming choice didn't make it obvious that part6 is the seventh. I think it
would be better to mention the label in the warning dialog, if there is a
label.

Similarly, when deleting an extended partition, the warning should make it very
clear that all the partitions contained in the extended partition will be
deleted too.

http://bugzilla.gnome.org/show_bug.cgi?id=576431

Signed-off-by: David Zeuthen <davidz@redhat.com>