Prepare project tree for port to udisks2
For now, nuke everything since

 - libgdu is essentially replaced with libudisks2

 - we don't want help files at all (help has been an empty doc
   since the start)

 - the nautilus formatting extension should probably be in Nautilus
   proper, not here

 - the notification daemon for SMART should probably be part of
   gnome-settings-daemon

basically leaving Palimpsest as the only thing we need here. And since
we want to move to using e.g. glade directly in the project, all that
code needs to be redone anyway. Of course we will still re-use a lot
of the existing code (it will be copy-pasted in), so this is far from
a total rewrite.

Signed-off-by: David Zeuthen <davidz@redhat.com>