Add "Disk Image Mounter" utility and associate with application/x-cd-image
The MINE association makes the utility show up in Nautilus for .iso
files... basically, if you open an ISO file (or even a selection of
multiple ISO files) with this tool through Nautilus the right thing
happens... that is, for each ISO file, a loop device is created and
its filesystem is mounted. This results in a set of GVolume/GMount
objects in Nautilus' sidebar and the shell will run its sniffing
machinery and ask if you want to e.g. watch the DVD movie.

Additionally, the loop.autoclear option is used so the loop device
goes away when the loop device is unmounted.

Signed-off-by: David Zeuthen <davidz@redhat.com>