Add a settings dialog for power management and acoustic levels
This feature is convenient if for always-on systems (for example
serving media files to a television or set-top-box) that are idle most
of the time. Reduces power usage and noise levels - save the planet!

We also show a ZZZ icon next to a drive if it is currently in standby
mode. We do this by polling all drives every few seconds - this is not
a problem as people don't generally have the Disks application running.

 http://people.freedesktop.org/~david/gnome-disks-settings-item-in-drive-menu.png
 http://people.freedesktop.org/~david/gnome-disks-settings-power-management.png
 http://people.freedesktop.org/~david/gnome-disks-settings-acoustic.png
 http://people.freedesktop.org/~david/gnome-disks-zzz-icon-for-sleeping-disks.png

Signed-off-by: David Zeuthen <zeuthen@gmail.com>