Don't show "standby timeout" setting for SSD drives
Some SSD drives (erroneously?) has support for the standby timeout
timer (hdparm -S) even though it doesn't make much sense as is
intended for spinning down drives.

So disable it for now. If it turns out that this is actually used for
something in SSDs we can bring it back...

Signed-off-by: David Zeuthen <zeuthen@gmail.com>