disk-image-mounter: leave mounting to gnome-shell
With GNOME Shell in GNOME 3.6, only devices mounted/unlocked by the
shell itself will result in notifications and addition of items to the
"Removable Devices" icon in the tray area. This is by design and it is
a good change - see

 https://bugzilla.gnome.org/show_bug.cgi?id=660595

for the rationale behind this.

The change is, however, at odds with how the Disk Image mounter since
attaching a disk image is very similar to attaching a USB device or
inserting an optical disc - we want the same user experience for disk
images. For example, double-clicking the file Fedora17.iso in the file
manager should be the same as attaching a USB stick branded "Fedora
17" (you should get a notification with "Open in Boxes" and "Eject"
buttons). Similarly, double-clicking the file "Armageddon.iso" and
inserting the DVD disc branded "Armageddeon" should result in the same
behavior (you get a notification saying "Open in Totem" and "Eject"
buttons).

After talking to Cosimo we decided to just change how GVfs works so
loop devices set up by the user itself (the Loop.SetupByUID property
from udisks is checked) gets the should_automount flag set to
TRUE. This change is here

 http://git.gnome.org/browse/gvfs/commit/?id=88ffd85e73fec8611386530962fb35451e28571c

This change makes gnome-disk-image-mounter a lot simpler since we
don't have to go through all partitions and mount everything. The
change also has the benefit that LUKS now work properly on disk
images.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>