Try opening the device file directly when creating images of CD/DVD discs
This avoids a polkit dialog for the common case where the user already
has an ACL on e.g. /dev/sr0. Also attempt to dynamically load the
libdvdcss library and use its interfaces if it's available (many other
apps include Brasero and Totem does the same thing). See
http://www.videolan.org/developers/libdvdcss.html for more information
about this library and what it does.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>