Add window title to Change Filesystem Label dialog
In GNOME 3.6 the window manager now displays window titles even if the
dialog is transient for its parent window (In GNOME 3.0 through 3.4
the WM displayed the window as a sheet - e.g. without the window
title). As no name has been set for the window, the window manager now
simply shows "gnome-disks" which is not very friendly. The fix is
simply to give the window a meaningful title.

Also remove the GtkLabel saying "Change Filesystem Label" since this
is now redundant with the window title

Note that this commit does not break string freeze since it re-uses
existing strings.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>