Add window title to the "Unlock LUKS" dialog
In GNOME 3.6 the window manager now displays window titles even if the
dialog is transient for its parent window (In GNOME 3.0 through 3.4
the WM displayed the window as a sheet - e.g. without the window
title). As no name has been set for the window, the window manager now
simply shows "gnome-disks" which is not very friendly. The fix is
simply to give the window a meaningful title.

Also remove the GtkLabel saying "Enter passphrase to unlock" since
this is now redundant with the window title

This commit does is re-using existing strings.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>