Use libdvdcss for creating disk images of DVDs, if available
Without this change, the archive process may fail for some DVD video
discs with "Add. Sense: Read of scrambled sector without
authentication".

This commit does NOT add a hard dependency on libdvdcss (which would
be bad as this library isn't even shipped in most OSes) since it's
accessed via dlopen(). If libdvdcss is not available on the system, we
simply fall back to reading data using the standard libc interfaces
(e.g. read(2)).

License-wise there is no problem with this approach: libdvdcss is
licensed under the GPLv2+ which is exactly the same license as Disks
itself.

This dlopen() approach is nothing new - it's also used in Brasero,
VLC, mplayer, Totem and many other applications.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>