Add ensure_unused() method to "shut down" a whole device
This is useful to avoid pointless error dialogs complaining e.g. that
there are mounted filesystems on the media being ejected or
formatted... so instead of presenting this such an error to the user,
this method will go ahead and unmount/lock all relevant
filesystems/LUKS devices for the device.

Many users expect this kind of behavior since they do not really keep
track of implementation details such as whether a device is mounted or
unlocked.

Upcoming commits will make use of this new infrastructure in several
places.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>