Detect and convey MD-RAID split-brain syndrome
See this udisks commit for more information about the syndrome

 http://cgit.freedesktop.org/udisks/commit/?id=301ab668ff65c4e9f28037251bedba26543b6024

The fact that Linux / mdadm allows you to end up in a split-brain
situation in the _first_ place is pretty horrific (that's the price
you pay for flexibility, loosely coupled systems etc. etc.).

Anyway, in Disks, we cope with the situation by conveying the device
files in question and a link to Wikipedia. It looks like this

 http://people.freedesktop.org/~david/gnome-disks-mdraid-split-brain.png

Ideally there would be a page on the Linux RAID wiki at

 https://raid.wiki.kernel.org/index.php/Linux_Raid

but that is not how things work right now.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>