Add a unit combo-box to the create partition dialog
When creating a new partition, we used to default to MB which would
yield very big numbers for normal-size disks (e.g. 1TB), for example
the user would have to enter 700000 to create a 700GB partition. Not
exactly user-friendly (#689362).

Instead, we now default to a better unit so numbers are 2-3 digits
(e.g. choose unit GB for a 1 TB disk). Additionally, we also provide a
combo box so the user can choose the unit himself. This also allows
finer control of how big of a partition to create which is something
that some users have requested (#643159).

https://bugzilla.gnome.org/show_bug.cgi?id=689362
https://bugzilla.gnome.org/show_bug.cgi?id=643159
http://people.freedesktop.org/~david/gnome-disks-create-partition-unit-combobox.png

Signed-off-by: David Zeuthen <zeuthen@gmail.com>