Show block device, even if size is zero
Why? Because under some circumstances the mdadm(8) command creates
"inactive" devices on failure paths - for example if asked to start an
array with a given UUID and not enough members are present, the
created device is left behind. I think this is a bug but instead of
second-guessing (e.g. stopping the inactive MD device), it's better to
just show the mess so the user/administrator himself can take action.

We do keep the "hide device of size zero" rule for loop devices since
normal Linux systems - for historical reasons - have them by default.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>