Avoid hiding widgets only to show them again
Doing so causes focus problems resulting in the inability to click a
button if, say, the window is continously updated (e.g. update_all()
called several times a second).

This problem manifested itself if the user was creating disk images
for two devices and the first finished before the second. Since the
second was still in progress the window was continously updated. This
resulted in an inability to click the eject button for the former.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>