Avoid needing gsd during build when --disable-gsd-plugin is used
PKG_CHECK_MODULES was previously unconditionally requiring
gnome-settings-daemon to be available, even when it was later
not going to be used because the --disable-gsd-plugin configure
switch was passed.

Move the check to only require g-s-d when it's actually going to be
used.

While at it, also fix the indentation of the msg_gsd_plugin=yes line.

https://bugzilla.gnome.org/show_bug.cgi?id=697146