GduVolumeGrid: GridElement should hold text, not markup
I see no reason why markup could ever possibly get into this variable,
so rename it to text and use the appropriate pango function for text.
This allows us to reuse this text for the grid's accessible name.

If somehow we wind up with markup appearing in the grid, then we'll need
to revisit this, but I think it's safe....

https://bugzilla.gnome.org/show_bug.cgi?id=747490