Use the first grid element for the initial accessible name of the grid
Instead of "Volumes Grid", so the user doesn't have to switch to another
grid element and back to read the first partition, and so that it is
possible to read the first partition when there is only one partition.

Also, use gdu_volume_grid_get_no_media_string() for the accessible name
when appropriate.

https://bugzilla.gnome.org/show_bug.cgi?id=747490