notify: Avoid failure on startup
With the new g-s-d architecture, we needed to start the helper as soon
as possible, but starting it as part of the initialisation of the
session meant that we would have needed to register with the session.

We don't need to start this early, and it's easier to start in the
background as an app, and not register.

https://bugzilla.gnome.org/show_bug.cgi?id=772863#c9