Add tooltip text to Auto-clear switch
Closing #769461 which points out that the auto-clear option for loop
devices is nowhere explained in the UI.
From its UDisks description: "If TRUE, the kernel will automatically clear
the loop device when the last closer closes the device. This typically
happens when the loop device is unmounted."
I chose the wording "Detach loop device after unmount action".