Add app menu entry to create an empty disk image
New images can be created with a given size in the selected folder
(being sparse files depends on the OS, works on Linux).
They will directly be attached as loop device for further
formatting/partitioning. The feature is useful in order to set up e.g.
encrypted LUKS containers which should be shared with other people.