Fix wrong behavior of filesystem creation dialog
As soon as a passphrase was entered the resulting filesystem was always
encrypted even if the selection was already changed to a non-luks option.
This is mainly due to the fact that luks+ext4 results in ext4 as fstype
and the decision to encrypt is just based on the passphrase length.

A new encryption property is therefore introduced to distinguish between
the cases instead of always clearing the passphrase entry in the dialog
as soon as the selection is changed. It should be better to handle and
pave the way for a split from the entangled LUKS+ext4 option.

https://bugzilla.gnome.org/show_bug.cgi?id=732566