Fix unmounting and probing of mounted volumes
Everytime all volumes of a device were to be umounted
(e.g. in order to eject the device) the LUKS partitions
have not been closed due to a bug. Also the benchmark
dialog preselected the write option for every device but
was supposed to do this only for unused devices.
This is because of broken code in gdu_utils_is_in_use_full
which always returned FALSE or TRUE depending on whether a
pointer was passed in filesystem_to_unmount_out and also its
code to set encrypted_to_lock_out was never executed then
and gdu_utils_is_in_use always returned FALSE.

After repairing the code the iterative umounting of all
mounted volumes works again and the write option in the
benchmark is only preselected for unmounted devices.

https://bugzilla.gnome.org/show_bug.cgi?id=780865