Disable auto-clear before operations needing unmount
For some actions the volume is ensured to be unmounted before
the action is started. The auto-clear of loop devices has to
be disabled because otherwise the whole device is gone after the
unmount and then the planned action fails with an error.

This is now handled by deactivating auto-clear if the last used
volume of a device is going to be unmounted.

https://bugzilla.gnome.org/show_bug.cgi?id=780865