Clarify meaning of encryption/mounting options
The dialogs to edit mount and encryption options
are basically a graphical user interface to add
corresponding entries in /etc/fstab and /etc/crypttab
but the scope of the words 'automatic' and 'startup'
is unclear.

It is more precise to relate to the automatic (now called default)
handling of the user session vs. manual policies for system startup.

https://bugzilla.gnome.org/show_bug.cgi?id=768769