Redesign Format Dialog
Since the format dialog used hiding of widgets it was
sometimes jumping around. The appearance of the passphrase
entries were bound to a special LUKS+Ext4 mixed entry
within a combo box. Also that apporach does not scale well
if other filesystems should be listed for encryption.

The dialog is redesinged and implemented with several pages,
and thus adaptive to the needs of adding partitions, a new
custom filesystem list and encryption, step for step.
The default FS choices are still kept and shown via radio
buttons on the format page.

https://bugzilla.gnome.org/show_bug.cgi?id=770738