Show mount options for swap volumes
The mount option dialog for editing /etc/fstab was
only shown for swap volumes if they were already
present in /etc/fstab.

Now the mount options are always accessible for
swap volumes. Their mount point is fixed to 'none'
because that is the only valid option, opts default
to 'sw' and type to swap. The irrelevant options
are insensitive.

https://bugzilla.gnome.org/show_bug.cgi?id=785453