Add unmount function to edit filesystem dialog
Currently, attempting to change the label of a mounted filesystem will
fail for most filesystem types, with no warning of this presented to the
user.

Change this so that the filesystem will be unmounted prior to changing
the label if it is currently mounted and not of type extY. A warning will
be presented to the user when this is to occur.

https://bugzilla.gnome.org/show_bug.cgi?id=784619