Correct text for partition size entry
The tooltip was always claiming to have MB as unit
while the unit could be changed to any other.
The new image dialog was talking about partitions.

Mentioning the unit is removed. In the label and tooltip
of the new image dialog "partition" is replaced by "image".
The tooltip markup text is deleted instead of updated
because it's not in use.

https://bugzilla.gnome.org/show_bug.cgi?id=786276