Keep FS mounted while growing partition (string freeze)
Growing the partition for offline resize caused
a rediscovery in UDisks to led to GVFs automount
which lets the Disks filesystem resize call fail.

A good solution is to keep the filesystem mounted
while growing the partition, as it is done in the
online resize already. This way there will be no
rediscovery and automount, and the filesystem can
be unmounted afterwards to be grown. From this point
there will be no automount coming because the partition
size is not touched anymore.

Same as on master except for only reusing translated
strings by taking shorter error messages.

https://bugzilla.gnome.org/show_bug.cgi?id=788326