restore: don't attempt to open target read-only
This code seems to be duplicated from gducreatediskimagedialog.c, but in
this context we need 'fd' to be writable. We could still attempt to
open() the device directly (with O_WRONLY) but as the comment says, by
definition /dev/sr* is read-only by nature (or at least, cannot be
written to in this way), so it is unlikely to be of any use.

In any case, in practice this code path will never be taken since the
"Restore" action is not available for optical media.

https://bugzilla.gnome.org/show_bug.cgi?id=792581