Fix unmount
Unmounting was not working because unuse_data_iterate() was trying
to unmount, then recurse back into itself, then check if it was
mounted again.  The second check would still see the fs as mounted
even though it had just been unmounted, so it would try to unmount
it again, which would fail.  Inhibit the second mounted check.

Closes: #101