build: Revise dependencies
gnome-disk-utility builds an internal static library called `libgdu`
that depends on a set of libraries (`gtk+-3.0`, `udisks2`, ...).
Then this static library is used when building `gnome-disks` and
`gnome-disk-image` mounter so the libraries exposed in `libgdu`'s
headers are indirect dependencies.

Only these indirect dependencies, not those used when building the
target itself, have been added to the `libgdu` internal dependency
so they are not necessary to be added explicitly on new targets.

The generated enums header has also been added to the dependency as
it is also an indirect dependency.