Wait until UDisks updates the unmounted filesystem
Before the Filesystem.Unmount method returns it does
not make sure that the mount point is removed from
the list stored as filesystem property.
I solved this once with a waiter in
https://github.com/storaged-project/udisks/pull/574
but since a couple of months the error is back.
Work around this in GNOME Disks by waiting
until the property has changed (i.e., the list has
a different size). Just in case something else gets
messed up, a timeout of 5 seconds is added, too.

In future UDisks releases this will be fixed by
https://github.com/storaged-project/udisks/pull/728