Always show other filesystem format choices
The decision to hide the option to use other filesystems than the standard
choices when invoked from Nautilus often was a source of confusion because
users had seen this option before when they were using GNOME Disks directly
or they were searching for the filesystem they needed but it was hidden and
they never considered opening GNOME Disks directly.

Always display the choice to select any other filesystem to resolve this
confusion. (This is particularly important when people have to use the exFAT
filesystem for, e.g., their camera SDXC cards. Currently I don't have such
a device to see how we could autodetect it and suggest it automatically
among the standard choices. But anyway, the other filesystems should be
available through Nautilus, too.)