Use special parameter when requesting the maximal partition size
This is a workaround and it would be nice to either handle this better
in libblockdev or by creating the partition and then shrinking it.
As having multiple GB partitions is common, I think the corner case can
be ignored for now but it should be documented in the code.

Fixes https://gitlab.gnome.org/GNOME/gnome-disk-utility/-/issues/138