Ignore the --xid option when formatting from Nautilus
The --format-device option required the --xid option to be set because
under X11 the idea was that GNOME Disks provides a dialog on top of
Nautilus. However, the integration wasn't really good because it lacked
any progress and error reporting. Under Wayland it was finally
completely broken to have GNOME Disks provide a dialog for Nautilus.
Instead, we opened the app in the regular way and displayed the dialog
there as usual which has the benefit of seeing the job state.
Do the same now under X11 by just ignoring the --xid option.