restore: do not wipe disk after error if it's a dismissed auth error
The wiping does not need privileges in contrast to opening the disk for
arbitrary operations. That means that the wiping that was done on error
was succeeding even though the user canceled the polkit auth process.
This is confusing and may lead to the user thinking that they could
accidentially wipe anything because it happens after canceling the auth
process - but it is also good to not wipe if the user changes the mind
after confirming the first dialog and it's nice to give a second chance
when the auth dialog comes.
Detect the dismissed auth error to distinguish it from other errors
happening during the operation and only wipe the disk after those
errors were encountered but not for the dismissed auth error coming
from canceling the polkit auth process.

Originally reported in
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=987671