Improve the visual appearance of the devtab toolbar
Improves the visual appearance of the Devtab Grid Toolbar by
making the toolbar a linked widget with the Devtab Grid above.
This prevents the odd visual look currently where the buttons in
the toolbar have squared-off corners on the insides, while
maintaining the UX-based separation of the 'delete-partition'
button (which is destructive) from the other buttons.