manager: Remove drive from local store iff it's a drive
We were removing matching (sub)object regardless of whether the item
was a drive or not, and this resulted in the item being removed
from our cache when it's partition table was removed.

Fix it by restricting the removal of an item if and only if it's
a UDisksDrive.

Fixes #327