benchmark-dialog: Cleanup stats formatting
For formatting transfer_rate we have a dedicated function but when formatting access time we format everything manually with almost the same steps as formatting transfer rate.

By simply modifying the format_transfer_rate to accept a boolean to differentate between access time and transfer rate. We can make it format all the stats, this reduces code repetition and makes the code cleaner.