gdu-block-row: Minor tweaks to the "Confirm Check" confirmation dialog
The "Confirm Check" title isn't phrased as a direct question. Also,
the "Ok" button isn't capitalized correctly and it isn't a standard
term to use on this button.

To fix this, rename the confirmation dialog to "Check Filesystem?" and
the "Ok" button to "Start Check" (and add a mnemonic to this button)

Closes https://gitlab.gnome.org/GNOME/gnome-disk-utility/-/issues/371