INSTALL: Document new invocations/workarounds for Flatpak development
The interactive Flatpak build from a checked out repo is a bit messy
because 'flatpak build' does not carry the build settings over from the
Flatpak JSON manifest. There also is a small bug in finding the linker
with gcc but with clang we can give the full path as workaround. And
actually the first issue to hit is that Flatpak derives the extension
ref from the GNOME SDK instead of the FreeDesktop SDK which is the
real base, and thus it looks for a 'master' FreeDesktop SDK extension
instead of the YEAR[beta] ref.