gdu-format-disk-dialog: Tweak the "Erase Existing Data" option
"Erase Existing Data" is misleading, because it can induce the
user to think that their data won't be affected if this option
is off. Also, this string doesn't properly explain that data
will be overwritten.

So, rename the option to "Overwrite Existing Data" and add an
info popover explaining about this option, just like in
GduFormatVolumeDialog.