Remove duplicate mime types to prefer new image mount UI
For a regular disk image there were two possible actions in Nautilus,
one is to mount the image, one is to write it out to a disk. The new
mount UI also offers the option to write to disk and therefore we can
use only this for the disk image mime types. The write action desktop
file stays around because it supports compressed disks which can't be
mounted directly.