block-row: Change layout of information grid
The current information grid is not mobile or translation friendly.
The labels can be arbitrarily sized, are placed side-by-side,
and do not wrap.

This commit changes the layout of the grid so that the type
of information is placed above the label, and makes all labels
wrap so that they account for translations or longer UUIDs,
sizes, etc.