handle extended partitions in drive view
extended partitions in drive view were not handled properly as they were
considered to be normal partitions taking up size on disk. Now, in case
of extended partitions, the end offset is not incremented beyond the
start of the extended partition.
Furthermore, a check if the extended partition has some free block after
all partitions are iterated through