Update doap file to add co-maintainer
Inam did a lot of things with the GTK4 port, addressing bugs from the
rework, MR reviews and more. He knows the code better than I do and
showed continued ownership so that it makes sense to be a co-maintainer
for MR reviews and merges, managing issues, pushing to branches and so
on.