benchmark-dialog: accomodate vertical and horizontal axis titles
To accomodate vertical and horizontal axis titles, reduced the size of
the graph and added graph size to graph_data which is updated wherever
we are drawing labels.

As we are reducing the size of graph, the card css style cannot be
applied directly so, I am now drawing a background manually to the
graph.