Install missing libgnome_disks.so
There were two issues. First, because of missing soname, the NEEDED path
was not the base name. So gnome-disks had to load the libgnome_disks.so from
the build directory. And meson is not capable to rewrite that path. However
it can rewrite RPATH instead. Second, the library was not installed at all.

Fixes #455